+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v2.3.0] - 2023-12-05
 
 ### Added
 
@@ -779,4 +779,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v2.1.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.0.1...v2.1.0
 [v2.1.1]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.1.0...v2.1.1
 [v2.2.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.1.1...v2.2.0
-[Unreleased]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.2.0...HEAD
+[v2.3.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.2.0...v2.3.0